Bag and Tag Rush!
Did not commit consistently during the rush leading up to bag and tag so
there is a large assortment of changes here. Bottom line is that this
code is sufficient to run the robot with very basic functionality.
Elevator Motion Magic is still in testing but for the most part looks
like all it is lacking is some proper tuning. A simple Percent Voltage
based command is included in this commit to run the elevator for
hardware testing.
As for
intake,
electrical
issues with a breakout board prevented us from testing the intake raise
motor, this will have to be tested on the practice robot.